+++ src/main/java/com/leonyk2/mcmod/item/ModCreativeModeTab.java
@@ -1,6 +1,7 @@
 package com.leonyk2.mcmod.item;
 
 import com.leonyk2.mcmod.McMod;
+import com.leonyk2.mcmod.block.ModBlocks;
 import net.minecraft.core.registries.Registries;
 import net.minecraft.network.chat.Component;
 import net.minecraft.world.item.CreativeModeTab;
@@ -29,6 +30,19 @@ public class ModCreativeModeTab {
                         pOutput.accept(ModItems.TAGE_WACH_MUSIC_DISC.get());
                         pOutput.accept(ModItems.TAGE_WACH_BB_MUSIC_DISC.get());
                         pOutput.accept(ModItems.LEICHE_MUSIC_DISC.get());
+
+                        pOutput.accept(ModBlocks.OAK_TABLE.get());
+                        pOutput.accept(ModBlocks.DARK_OAK_TABLE.get());
+                        pOutput.accept(ModBlocks.SPRUCE_TABLE.get());
+                        pOutput.accept(ModBlocks.BIRCH_TABLE.get());
+                        pOutput.accept(ModBlocks.JUNGLE_TABLE.get());
+                        pOutput.accept(ModBlocks.ACACIA_TABLE.get());
+                        pOutput.accept(ModBlocks.MANGROVE_TABLE.get());
+                        pOutput.accept(ModBlocks.CHERRY_TABLE.get());
+
+                        //pOutput.accept(ModItems.TEST_PONDER.get());
+
+                        //pOutput.accept(ModBlocks.TEST_PONDER_ENTRY);
                     }))
                     .build());