+++ src/main/java/com/leonyk2/mcmod/item/ModFoods.java
@@ -1,5 +1,6 @@
 package com.leonyk2.mcmod.item;
 
+import net.minecraft.world.effect.MobEffectInstance;
 import net.minecraft.world.food.FoodProperties;
 
 public class ModFoods {